fix bugs

+++ server/src/api/services/cart/index.js
@@ -24,16 +24,14 @@ const addToCart = async ({ userId, product = {}, quantity }) => {
 
 const updateCart = async ({ userId, product = {}, quantity }) => {
     try {
-        // const foundProduct = await getProduct(product.id);
-        // if (!foundProduct) {
-        //     throw createError.NotFound('Product not found!');
-        // }
+        const foundProduct = await getProduct(product.id);
+        if (!foundProduct) {
+            throw createError.NotFound('Product not found!');
+        }
 
-        // if (foundProduct.shopId !== product.shopId) {
-        //     console.log(foundProduct);
-        //     console.log(product.shopId);
-        //     throw createError.NotFound('Product not found!');
-        // }
+        if (foundProduct.shopId !== product.shopId) {
+            throw createError.NotFound('Product not found!');
+        }
 
         if (quantity === 0) {
             return await deleteCartItem({ userId, product });
@@ -83,10 +81,19 @@ const getUserCart = async ({ userId }) => {
     }
 };
 
+const getCartLength = async ({ userId }) => {
+    try {
+        return await redisClient.hLen(`cart:${userId}`);
+    } catch (error) {
+        throw error;
+    }
+};
+
 module.exports = {
     addToCart,
     updateCart,
     deleteCart,
     getUserCart,
     deleteCartItem,
+    getCartLength,
 };
